% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_rank.R
\name{boot_rank}
\alias{boot_rank}
\title{Calculating bootstrapped model coefficients for a bradley terry rank model. Bootstrap based on ID column rather than individual rows.}
\usage{
boot_rank(data, IDcol, formula, nboot = 100, diagnostics = TRUE,
  CIplot = TRUE, level = 0.95)
}
\arguments{
\item{data}{original data frame}

\item{IDcol}{ID column within data frame}

\item{formula}{model formula}

\item{nboot}{number of bootstraps to run}

\item{diagnostics}{TRUE/FALSE. Produce diagnostic violin plots of coefficients}

\item{CIplot}{TRUE/FALSE. Produce CI plots of coefficients.}

\item{level}{confidence level for CI calculations}
}
\description{
Calculating bootstrapped model coefficients for a bradley terry rank model. Bootstrap based on ID column rather than individual rows.
}
\keyword{bootstrap}
\keyword{bradley}
\keyword{rank}
\keyword{terry}
