% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBoot.R
\name{CIboot}
\alias{CIboot}
\title{Calculating confidence intervals around bootstrap estimates from boot_rank}
\usage{
CIboot(boots, CIMethod, level)
}
\arguments{
\item{boots}{output from boot_res of bootstrapped values}

\item{CIMethod}{confidence interval calculation method. "quantile" for using tails of 
empirical distribution for CI and getting estimate from median value; "SD" for using standard deviation of distribution;
"qvcalc" for using quasivariance (and this getting interval around reference
level as well)}

\item{level}{confidence level for CI calculations}
}
\description{
Calculating confidence intervals around bootstrap estimates from boot_rank
}
\keyword{bootstrap}
\keyword{bradley}
\keyword{rank}
\keyword{terry}
