% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paircomp.R
\name{paircomp}
\alias{paircomp}
\title{Take long format rank data and calculate pairwise wide data frame with one row per comparison and one column per option}
\usage{
paircomp(options, ranks, IDs, ties = "half", fixnames = TRUE)
}
\arguments{
\item{options}{vector of options}

\item{ranks}{vector of ranks}

\item{IDs}{vector of IDs}

\item{ties}{method for solving ties. does nothing at the moment but might do later.}

\item{fixnames}{make names of options syntactically valid in case of spaces or punctuations in the name}
}
\description{
Take long format rank data and calculate pairwise wide data frame with one row per comparison and one column per option
}
\examples{

#write function to take data and make pairwise comparison matrix
}
\keyword{comparisons}
\keyword{matrix}
\keyword{pairwise}
\keyword{ranks}
